# specs/001-intuitive-data-management/contracts/api-contracts.yaml
openapi: 3.0.3
info:
  title: ChurnVision Data Management API
  description: Intuitive data management endpoints for non-technical users
  version: 1.0.0
  contact:
    name: ChurnVision Enterprise
    url: https://churnvision.example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: License-validated JWT token
  schemas:
    Employee:
      type: object
      required: [hr_code, full_name, position, structure_name, status]
      properties:
        hr_code:
          type: string
          description: Unique employee identifier within dataset
          example: "EMP001"
        full_name:
          type: string
          minLength: 2
          maxLength: 100
          description: Employee's complete name
          example: "John Smith"
        position:
          type: string
          description: Job title/role
          example: "Senior Software Engineer"
        structure_name:
          type: string
          description: Department/organizational unit
          example: "Engineering"
        status:
          type: string
          enum: [active, inactive, terminated]
          description: Employment status
          example: "active"
        last_modified:
          type: string
          format: date-time
          description: Last modification timestamp
        modified_by:
          type: string
          description: User ID who made last change
        data_quality_score:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: Data completeness score
          example: 0.95
    EmployeeList:
      type: object
      required: [data, total, page, page_size, has_more]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        total:
          type: integer
          minimum: 0
          description: Total number of employees matching criteria
        page:
          type: integer
          minimum: 0
          description: Current page number (0-based)
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
        has_more:
          type: boolean
          description: Whether more pages are available
...
    EmployeeUpdate:
      type: object
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 100
        position:
          type: string
        structure_name:
          type: string
        status:
          type: string
          enum: [active, inactive, terminated]
        change_reason:
          type: string
          description: Optional reason for the change
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
security:
  - BearerAuth: []
paths:
  /data-management/employees:
    get:
      tags: [Data Overview]
      summary: Get paginated employee list with filtering
      description: Retrieve employees for data overview and navigation (User Story 1)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-based)
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Number of items per page
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, terminated]
          description: Filter by employment status
        - name: structure_name
          in: query
          schema:
            type: string
          description: Filter by department/structure
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [full_name, position, structure_name, last_modified]
            default: full_name
          description: Sort field
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction
      responses:
        '200':
          description: Employee list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeList'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /data-management/employees/{hr_code}:
    get:
      tags: [Employee Details]
      summary: Get individual employee details
      description: Retrieve specific employee information for editing
      parameters:
        - name: hr_code
          in: path
          required: true
          schema:
            type: string
          description: Employee HR code
      responses:
        '200':
          description: Employee details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [Employee Editing]
      summary: Update employee information
      description: Edit employee data using simple form interface (User Story 3)
      parameters:
        - name: hr_code
          in: path
          required: true
          schema:
            type: string
          description: Employee HR code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...